+++ packages/schemas/src/entitites/index.ts
@@ -3,5 +3,6 @@ export * from './cards';
 export * from './enums';
 export * from './infographics';
 export * from './items';
+export * from './rounds';
 export * from './steps';
 export * from './users';
