+++ packages/schemas/src/round.ts
@@ -1,17 +1,5 @@
 import { z } from 'zod';
 
-export const roundLinkSchema = z.string().nullish();
-
-export const roundSchema = z.object({
-  id: z.number(),
-  link1: roundLinkSchema,
-  link2: roundLinkSchema,
-  createdAt: z.string().datetime().nullish(),
-  updatedAt: z.string().datetime().nullish(),
-});
-
-export type Round = z.infer<typeof roundSchema>;
-
 export const updateRoundLinkBodySchema = z.object({
   link: z.string().optional(),
 });
