+++ packages/web/src/types/rounds.ts
@@ -4,5 +4,7 @@ import { z } from 'zod';
 
 export const completeRoundSchema = roundSchema.extend({
   categories: z.array(categorySchema),
+  createdAt: z.string().datetime().nullish(),
+  updatedAt: z.string().datetime().nullish(),
 });
 export type CompleteRound = z.infer<typeof completeRoundSchema>;
